Fix up paths from rename
