+++ custom_components/simply_magic_areas/tests/conftest.py
@@ -6,7 +6,7 @@ import logging
 import pytest
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
-from custom_components.magic_areas.const import (
+from custom_components.simply_magic_areas.const import (
     AREA_TYPE_INTERIOR,
     CONF_CLEAR_TIMEOUT,
     CONF_EXTENDED_TIMEOUT,
