+++ custom_components/simply_magic_areas/tests/test_init.py
@@ -4,7 +4,7 @@ import logging
 
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
-from custom_components.magic_areas.const import DOMAIN
+from custom_components.simply_magic_areas.const import DOMAIN
 from homeassistant.components.light import DOMAIN as LIGHT_DOMAIN
 from homeassistant.components.select import DOMAIN as SELECT_DOMAIN
 from homeassistant.components.switch import DOMAIN as SWITCH_DOMAIN
